refactor: 채팅 입력 관련 코드 useChatInput hook으로 분리
