+++ src/components/Chat/Chatting/Chatting.tsx
@@ -1,4 +1,4 @@
-import { ChangeEvent, useRef, useState, useEffect, KeyboardEvent } from 'react';
+import { useRef, useState, useEffect } from 'react';
 import LatestMessageBox from '@components/Chat/LatestMessageBox/LatestMessageBox';
 import MyMessageBox from '@components/Chat/MyMessageBox/MyMessageBox';
 import OtherMessageBox from '@components/Chat/OtherMessageBox/OtherMessageBox';
@@ -6,13 +6,12 @@ import { Button } from '@components/common/Button/Button';
 import Flex from '@components/common/Flex/Flex';
 import IconButton from '@components/common/IconButton/IconButton';
 import Text from '@components/common/Text/Text';
-import useFileUpload from '@hooks/common/useFileUpload';
+import { useChatInput } from '@hooks/chatting/useChatInput';
 import useChatMessageQuery from '@hooks/queries/chat/useChatMesaageQuery';
 import { queryClient } from '@hooks/queries/common/queryClient';
 import useUserStatusQuery from '@hooks/queries/useUserStatusQuery';
 import { StompSubscription } from '@stomp/stompjs';
 import useSocketStore from '@stores/socketStore';
-import useToastStore from '@stores/toastStore';
 import { getFormattedDate } from '@utils/getFormattedDate';
 import { END_POINTS } from '@constants/api';
 import { CHAT_AUTO_SCROLL_LIMIT } from '@constants/size';
@@ -27,23 +26,32 @@ const Chatting = (props: ChattingProps) => {
 	const { userStatus } = useUserStatusQuery();
 	const { selectedChat } = props;
 	const { stompClient } = useSocketStore();
-	const { makeToast } = useToastStore();
+
+	const {
+		chatMessage,
+		inputImageRef,
+		inputFileRef,
+		messageEndRef,
+		handleMessageChange,
+		handleText,
+		handleImageUploadClick,
+		handleFileUploadClick,
+		handleImageChange,
+		handleFileChange,
+		handleKeyDown,
+	} = useChatInput(selectedChat, userStatus);
 
 	const { messages, fetchNextPage, hasNextPage, isFetching } =
 		useChatMessageQuery(selectedChat, userStatus?.profile.lastSeenTeamspaceId);
 
 	const [chatHistory, setChatHistory] = useState<ChatData | null>(null);
-	const [chatMessage, setChatMessage] = useState('');
 	const [prevHeight, setPrevHeight] = useState(0);
 	const [isScrollAtBottom, setIsScrollAtBottom] = useState(false);
 	const [isLatestMessageVisible, setIsLatestMessageVisible] = useState(false);
 	const [initialLoad, setInitialLoad] = useState(true);
 
 	const chatRef = useRef<HTMLDivElement>(null);
-	const messageEndRef = useRef<HTMLInputElement | null>(null);
 	const chatSubscribeRef = useRef<StompSubscription | null>(null);
-	const inputImageRef = useRef<HTMLInputElement | null>(null);
-	const inputFileRef = useRef<HTMLInputElement | null>(null);
 
 	useEffect(() => {
 		if (selectedChat && userStatus) {
@@ -190,138 +198,6 @@ const Chatting = (props: ChattingProps) => {
 			scrollElement?.removeEventListener('scroll', handleScroll);
 		};
 	}, []);
-
-	const handleMessageChange = (e: ChangeEvent<HTMLTextAreaElement>) => {
-		const { value } = e.target;
-		if (value.length <= 1000) setChatMessage(value);
-	};
-
-	const handleText = () => {
-		if (userStatus) {
-			stompClient?.send(
-				END_POINTS.SEND_MESSAGE(
-					userStatus.profile.lastSeenTeamspaceId,
-					selectedChat
-				),
-				{},
-				JSON.stringify({
-					chatType: 'TEXT',
-					content: chatMessage,
-					images: [],
-					attachments: [],
-				})
-			);
-		}
-
-		setChatMessage('');
-	};
-
-	const handleImageUploadClick = () => {
-		inputImageRef.current?.click();
-	};
-
-	const handleFileUploadClick = () => {
-		inputFileRef.current?.click();
-	};
-	const { isFileSizeExceedLimit, uploadFiles } = useFileUpload();
-
-	const handleImageChange = async (
-		event: React.ChangeEvent<HTMLInputElement>
-	) => {
-		if (event.target.files && event.target.files[0]) {
-			if (isFileSizeExceedLimit(event.target.files[0])) {
-				makeToast('이미지 크기는 최대 100MB입니다.', 'Warning');
-				return;
-			}
-			if (inputImageRef.current?.files) {
-				const imageUrl = await uploadFiles(
-					inputImageRef.current?.files,
-					'TEAMSPACE',
-					userStatus?.profile.lastSeenTeamspaceId
-				);
-				if (imageUrl && userStatus) {
-					stompClient?.send(
-						END_POINTS.SEND_MESSAGE(
-							userStatus.profile.lastSeenTeamspaceId,
-							selectedChat
-						),
-						{},
-						JSON.stringify({
-							chatType: 'IMAGE',
-							content: null,
-							images: [
-								{
-									name: inputImageRef.current?.files[0].name,
-									fileUrl: imageUrl[0],
-									size: inputImageRef.current?.files[0].size,
-								},
-							],
-							attachments: [],
-						})
-					);
-				}
-
-				setTimeout(() => {
-					messageEndRef.current?.scrollIntoView({ behavior: 'smooth' });
-				}, 500);
-			}
-		}
-	};
-
-	const handleFileChange = async (event: ChangeEvent<HTMLInputElement>) => {
-		if (event.target.files && event.target.files[0]) {
-			if (isFileSizeExceedLimit(event.target.files[0])) {
-				makeToast('파일 크기는 최대 100MB입니다.', 'Warning');
-				return;
-			}
-
-			if (inputFileRef.current?.files) {
-				const fileUrl = await uploadFiles(
-					inputFileRef.current?.files,
-					'TEAMSPACE',
-					userStatus?.profile.lastSeenTeamspaceId
-				);
-				if (fileUrl && userStatus) {
-					stompClient?.send(
-						END_POINTS.SEND_MESSAGE(
-							userStatus.profile.lastSeenTeamspaceId,
-							selectedChat
-						),
-						{},
-						JSON.stringify({
-							chatType: 'FILE',
-							content: null,
-							images: [],
-							attachments: [
-								{
-									name: inputFileRef.current?.files[0].name,
-									fileUrl: fileUrl[0],
-									size: inputFileRef.current?.files[0].size,
-								},
-							],
-						})
-					);
-				}
-
-				setTimeout(() => {
-					messageEndRef.current?.scrollIntoView({ behavior: 'smooth' });
-				}, 500);
-			}
-		}
-	};
-
-	const handleKeyDown = (event: KeyboardEvent<HTMLTextAreaElement>) => {
-		if (event.nativeEvent.isComposing) return;
-
-		if (event.key === 'Enter' && event.shiftKey) {
-			event.preventDefault();
-			setChatMessage((prev) => `${prev}\n`);
-		} else if (event.key === 'Enter' && !event.shiftKey) {
-			event.preventDefault();
-
-			if (chatMessage.trim().length > 0) handleText();
-		}
-	};
 
 	const handleLatestMessageClick = () => {
 		setIsLatestMessageVisible(false);
